Various fixes:
Cleaned the debug output of the threat maps:
modified trunk/ctp2_code/ai/mapanalysis/mapgrid.h

Made the game compile on Visual C++ 9.0:
modified trunk/ctp2_code/ai/strategy/goals/ctpgoal.cpp
modified trunk/ctp2_code/ai/strategy/goals/Goal.cpp
modified trunk/ctp2_code/ui/aui_ctp2/SelItem.cpp

Fixed an access of invalid memory:
modified trunk/ctp2_code/gs/gameobj/Order.cpp
modified trunk/ctp2_code/gs/gameobj/Unit.h

Fixed bug causing the game to crash if you reload a save game and an army is killed that has not had its turn:
modified trunk/ctp2_code/ai/strategy/scheduler/scheduler.cpp

Fixed some typos:
modified trunk/doc/user/playtest/Apolyton_README.txt


git-svn-id: http://ctp2.darkdust.net/anonsvn/trunk@876 34d2fc0c-33df-0310-8a34-85f8040ba873

+++ ctp2_code/ai/strategy/goals/ctpgoal.cpp
@@ -765,7 +765,7 @@ Utility CTPGoal::Compute_Matching_Value( const Agent_ptr agent ) const
 
 	sint32 squared_distance = 0;
 	double eta = ctpagent_ptr->GetRoundsPrecise(dest_pos, squared_distance);
-	double cell_dist = sqrt(squared_distance);
+	double cell_dist = sqrt(static_cast<double>(squared_distance));
 
 	Utility raw_priority = Get_Raw_Priority();
